refactor(TranslationSidebar): simplify component by utilizing master hook for state management and operations, removing redundant local state and imports
