+++ frontend/src/components/TranslationSidebar/hooks/index.ts
@@ -1,0 +1,12 @@
+// Export all hooks for easy importing
+export { useTranslationOperations } from './useTranslationOperations';
+export { useTranslationResults } from './useTranslationResults';
+export { useTextSelection } from './useTextSelection';
+export { useCopyOperations } from './useCopyOperations';
+export { useGlossaryOperations } from './useGlossaryOperations';
+export { useStandardizationOperations } from './useStandardizationOperations';
+export { useTranslationSidebarOperations } from './useTranslationSidebarOperations';
+
+// Export types for external use
+export type { TranslationConfig, TranslationResult } from './useTranslationOperations';
+export type { GlossaryTerm, GlossaryEvent } from './useGlossaryOperations';
